More renames
Renames so import of fortio/http does not conflict with net/http
Also cleanup top dir a bit
