+++ fgrpc/grpcrunner.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package grpc // import "istio.io/fortio/grpc"
+package fgrpc // import "istio.io/fortio/fgrpc"
 
 import (
 	"context"
