+++ fhttp/http.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package http // import "istio.io/fortio/http"
+package fhttp // import "istio.io/fortio/fhttp"
 
 import (
 	"bytes"
